move readers.i3s and readers.slpk to always-on readers
