+++ io/SlpkReader.hpp
