+++ io/private/esri/EsriUtil.cpp
