+++ io/private/esri/EsriUtil.hpp
