+++ io/private/esri/Obb.cpp
