+++ io/private/esri/Obb.hpp
