+++ test/unit/io/SlpkReaderTest.cpp
